refactor: issue options items 객체 파일 분리
